+++ fdb-record-layer-core/src/main/java/com/apple/foundationdb/record/query/plan/RecordQueryPlanner.java
@@ -2986,9 +2986,8 @@ public class RecordQueryPlanner implements QueryPlanner {
             @Nullable ScoredMatch scoredMatch = maybeSortedMatch.apply(null);
             if (scoredMatch != null) {
                 ComparisonRanges nextComparisonRanges = scoredMatch.getComparisonRanges();
-                if (!comparisons.isEqualities() && nextComparisonRanges.getEqualitiesSize() > 0) {
-                    throw new Query.InvalidExpressionException(
-                            "Two nested fields in the same and clause, combine them into one");
+                if (!comparisons.isUncommitedComparisonRangesEqualities() && nextComparisonRanges.getEqualitiesSize() > 0) {
+                    throw new Query.InvalidExpressionException("Two nested fields in the same and clause, combine them into one");
                 } else {
                     // nextComparisonRanges should not be null at this point
                     Objects.requireNonNull(nextComparisonRanges);